Extract ClasspathContainerState from ClasspathComputer
With the enhanced storage/save of the classpath container state now the
ClasspathComputer becomes more and more complex and contains too much
different tasks.

This extracts from the ClasspathComputer the parts that handle the state
into new class ClasspathContainerState to make it more maintainable.
